EVERYTHING SHOULD WORK in this commit. there were some serious problems in previous versions, espeically in GMRES

+++ gmres.h
@@ -4,7 +4,7 @@
 
 class GMRES : public Linear_Solver {
  public:
-  GMRES(const double tol, const double kmax);
+  GMRES(const double tol, const int kmax);
   ~GMRES(){}
   Eigen::VectorXd solve_linear_system(const Eigen::MatrixXd& A, const Eigen::VectorXd& b, const Eigen::VectorXd& x0) const;
   Eigen::VectorXd solve_linear_system(Eigen::VectorXd (*F)(const Eigen::VectorXd&), const Eigen::VectorXd& x, const Eigen::VectorXd& x0, const double dx) const;
